refactor(internal): move dev-only schema tooling under rbx/box/internal
The schema exporter and site builder are about developing rbx itself, not
about rbx functionality: nothing in the CLI reaches them, only the docs
build, the release tasks and CI. schema_urls stays in rbx/box because it is
runtime -- model_to_yaml and rbx fix call it to stamp headers.

The schemas reference page moves to a hidden "Internal reference" section:
it still builds and is linkable at /internal/schemas/, but is kept out of
the nav until there is more worth showing there.
