+++ docs/internal/schemas.md
