+++ rbx/box/internal/__init__.py
@@ -1,0 +1,6 @@
+"""Tooling for developing rbx itself, rather than rbx functionality.
+
+Nothing here is reachable from the CLI: these modules exist to build and
+publish artifacts about rbx (its JSON schemas, and whatever else follows) and
+are driven by the docs build, the release tasks, and CI.
+"""
