+++ tests/rbx/box/internal/test_schema_export.py
@@ -4,7 +4,7 @@ import json
 
 import pytest
 
-from rbx.box import schema_export
+from rbx.box.internal import schema_export
 
 
 def test_exports_every_documented_model(tmp_path):
@@ -43,7 +43,7 @@ def test_exported_schema_is_valid_json_and_relaxed(tmp_path):
 
 class TestSchemaPublish:
     def test_publish_layout(self, tmp_path):
-        from rbx.box import schema_publish
+        from rbx.box.internal import schema_publish
 
         schema_publish.build_site(tmp_path, version='1.4.2')
 
@@ -58,7 +58,7 @@ class TestSchemaPublish:
         assert 'Package' in index['models']
 
     def test_publish_merges_with_existing_versions(self, tmp_path):
-        from rbx.box import schema_publish
+        from rbx.box.internal import schema_publish
 
         schema_publish.build_site(tmp_path, version='1.3.0')
         schema_publish.build_site(tmp_path, version='1.4.0')
@@ -70,7 +70,7 @@ class TestSchemaPublish:
         assert (tmp_path / '1.3' / 'Package.json').is_file()
 
     def test_republishing_older_patch_does_not_demote_latest(self, tmp_path):
-        from rbx.box import schema_publish
+        from rbx.box.internal import schema_publish
 
         schema_publish.build_site(tmp_path, version='1.4.0')
         schema_publish.build_site(tmp_path, version='1.3.9')
@@ -80,7 +80,7 @@ class TestSchemaPublish:
         assert index['latest'] == '1.4'
 
     def test_versions_sort_numerically_not_lexicographically(self, tmp_path):
-        from rbx.box import schema_publish
+        from rbx.box.internal import schema_publish
 
         for version in ('1.9.0', '1.10.0', '2.0.0'):
             schema_publish.build_site(tmp_path, version=version)
@@ -91,7 +91,7 @@ class TestSchemaPublish:
         assert index['latest'] == '2.0'
 
     def test_ignores_unrelated_directories(self, tmp_path):
-        from rbx.box import schema_publish
+        from rbx.box.internal import schema_publish
 
         (tmp_path / '.git').mkdir()
         (tmp_path / 'assets').mkdir()
